Organise profile contributions by article
The profile of people with many contributions looked messy. This commit
tries to remedy this by grouping article contributions by article, and
hiding everything but the article name by default. There was also an
inconsistency in header tag usage on the profile page, which has been
fixed in this commit.

This fixes #209.